release 2.0

+++ src/main/java/com/ssafy/solvedpick/members/dto/UserInfoResponse.java
@@ -7,8 +7,9 @@ import lombok.*;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 public class UserInfoResponse {
+
     private String username;
-    private int point;
+    private Long point;
     private String tier;
     private int solvedCount;
     private int streak;
